refactor(traceroute): extract analytics into traceroute_analytics app
Move Neo4j, reach/coverage, stats/heatmap API views, analytics Celery impl, and management commands to traceroute_analytics. Keep /api/traceroutes/ routes and Celery task names via traceroute.tasks wrappers. Part of #247.

Made-with: Cursor
