+++ Meshflow/traceroute_analytics/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class TracerouteAnalyticsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "traceroute_analytics"
+    verbose_name = "Traceroute analytics"
